requirements.txt
langgraph~=0.4.8
beautifulsoup4
playwright~=1.52.0
langchain~=0.3.25
python-dotenv~=1.1.0
langchain-core~=0.3.64
langchain_openai~=0.3.27
langchain-community~=0.3.24
furl~=2.1.3
pydantic~=2.11.5
langchain-text-splitters~=0.3.8
html2text~=2025.4.15